(fix) Implement lambdas using the nodes array.
This was sooo much simpler. No extra logic for lambdas but some constraints.
Please use the functions exported as `Graph.Lambda` to operator on lambda nodes
they are special in the sense that they are atomics and have children. They have no
edges an exactly one child. The method `Lambda.createLambda` will take care of
the details.