(feature) Add dot output when an error occured.
Added it to topological sorting. To prevent flooding the terminal
it will be sent to fd3. You can get the output via `3> ...`. You
can print the graphs via `... 3> (> terminal-graphs --dot)` if you
have terminal-graphs installed globally. If you do not specify how
to use the fd3 output it will not be printed.