(fix) Add inside compounds calls callback correctly.
Previously functions that add nodes inside a compound did not pass the
newest graph to the callback. Instead the inner compound was passed, i.e.

```
Graph.addNodeIn(parent, n, graph, (newNode, newGraph) => newGraph)
```

in this example `newGraph` was not the graph after adding the node (as it
should be), but the updated `parent` node.